# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "API para gerenciamento de produtos.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lista todos as categorias",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Lista de todos as categorias"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Cria uma nova categoria",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição da categoria",
                                        "type": "string"
                                    },
                                    "img_path": {
                                        "description": "Caminho da imagem da categoria",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Exibe uma categoria especificada",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria encontrada"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Atualiza uma categoria especificada",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição da categoria",
                                        "type": "string"
                                    },
                                    "img_path": {
                                        "description": "Caminho da imagem da categoria",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Remove uma categoria especifica",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria removida com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista todos os produtos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista de todos os produtos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um novo produto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "min_sell_price",
                                    "cost_price",
                                    "barcode",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do produto",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição do produto",
                                        "type": "string"
                                    },
                                    "min_sell_price": {
                                        "description": "Preço mínimo de venda",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cost_price": {
                                        "description": "Preço de custo",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "barcode": {
                                        "description": "Código de barras do produto",
                                        "type": "string"
                                    },
                                    "weight": {
                                        "description": "Peso do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "length": {
                                        "description": "Comprimento do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "width": {
                                        "description": "Largura do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "height": {
                                        "description": "Altura do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "description": "Status do produto",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Cor do produto",
                                        "type": "string"
                                    },
                                    "size": {
                                        "description": "Tamanho do produto",
                                        "type": "string"
                                    },
                                    "material": {
                                        "description": "Material do produto",
                                        "type": "string"
                                    },
                                    "img_path": {
                                        "description": "Caminho da imagem do produto",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Exibe um produto especificado",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto especificado",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "min_sell_price",
                                    "cost_price",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do produto",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição do produto",
                                        "type": "string"
                                    },
                                    "min_sell_price": {
                                        "description": "Preço mínimo de venda",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cost_price": {
                                        "description": "Preço de custo",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "barcode": {
                                        "description": "Código de barras do produto",
                                        "type": "string"
                                    },
                                    "weight": {
                                        "description": "Peso do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "length": {
                                        "description": "Comprimento do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "width": {
                                        "description": "Largura do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "height": {
                                        "description": "Altura do produto",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "description": "Status do produto",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Cor do produto",
                                        "type": "string"
                                    },
                                    "size": {
                                        "description": "Tamanho do produto",
                                        "type": "string"
                                    },
                                    "material": {
                                        "description": "Material do produto",
                                        "type": "string"
                                    },
                                    "img_path": {
                                        "description": "Caminho da imagem do produto",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remove um produto especificado",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto removido com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}